% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualiza_dados.R
\name{atualiza_dados}
\alias{atualiza_dados}
\title{Atualiza dados do Brasileirão}
\usage{
atualiza_dados(project_id)
}
\arguments{
\item{project_id}{ID do seu projeto do Google Cloud. Ele deve possuir algumas autorizações específicas para conseguir importar os dados. Para mais informações, ver referências.}
}
\description{
Atualiza o `data.frame brasileirao` com os dados mais recentes
 disponíveis na fonte.
}
\references{
https://basedosdados.org/dataset/c861330e-bca2-474d-9073-bc70744a1b23?table=18835b0d-233e-4857-b454-1fa34a81b4fa
}
