% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{grafico}
\alias{grafico}
\title{Criador de gráficos com ggplot2}
\usage{
grafico(
  dados,
  x = NULL,
  y = NULL,
  color = NULL,
  fill = NULL,
  group = NULL,
  tipo = "hist",
  tema = 1
)
}
\arguments{
\item{dados}{Objeto do tipo `data.frame` que será utilizado para a criação do gráfico}

\item{x}{Variável que ficará no eixo x}

\item{y}{Variável que ficará no eixo y}

\item{color}{Variável que será utilizada na estética `color`}

\item{fill}{Variável que será utilizada na estética `fill`}

\item{group}{Variável que será utilizada na separação dos gráficos na função `facet_wrap`}

\item{tipo}{Tipo de gráfico que se deseja criar. Estão implementados os seguintes:
\describe{
  \item{pontos}{Gráfico de pontos}
  \item{hist}{Histograma}
  \item{dens}{Gráfico de densidade}
  \item{bar}{Gráfico de barras}
}}
}
\description{
Esta função cria gráficos usando o pacote `ggplot2` de uma maneira mais
simples. Além disso, é possível escolher o tema de interesse.
}
